Support for "-enable-iam-login" cloudsql-proxy arg
This adds support for setting the "-enable-iam-login" argument for
the cloudsql-proxy. This enables Postgresql database auth using the
Workload Identity SA for the application. Note that the sqluser must be
created with "type: CLOUD_IAM_SERVICE_ACCOUNT" and the username will be
the Workload Identity SA without the .gserviceaccount.com suffix.