Add support for omitting cloudsql-proxy sidecar
Setting the `useSqlProxy` option to `false` will skip the addition of the cloudsql-proxy, but will keep the creation
of the `ServiceEntry` making it possible to connect to the Cloud SQL database. The value defaults to `true` if not
set, for backward compatibility.

If the developers opt to use IamAuthentication to the Cloud SQL database the cloudsql-proxy is not needed.
The reason for using IamAuthentication is to simplify setup by using IAM instead of username and password.

Docs on using IamAuthentication via the connectors from Google directly in-process:
https://cloud.google.com/sql/docs/postgres/iam-logins#python

Also, reverted the change introduced in v2.1.0 as it seems the use case for this setting in the cloudsql-proxy could
is intended for connecting from your local machine (a database tool for instance)