Limit ServiceEntry to the namespace it is defined
A ServiceEntry should only take effect in the namespace where
it is defined.

This fixes an issue that a ServiceEntry from one namespace can
interfere with the traffic of another namespace.

The "exportTo" configuration with a value of "." defines an export to
the same namespace that the service is declared in.

Tests added.

Internal ref: [UTVIK-99]