Refactor istio-app-to-app
Split prometheus rule to own file, such that it can be tested and treated independently of other rules.
Rename "serviceAccounts" -> "from", and allow either serviceAccount or namespace and appname as parameter to define the policies.
Create one document for each rule. This allows for extending further if we want to add fine grained access as well.