Fix RequestAuthentication was not own document
If there was multiple elements in the ingress.internal list and one of them contained a jwt, the document would not ble split (i.e. --- was missing). This commit fix that, and also updates test to use the correct index now that the RequestAuthentication correctly is its own element

Internal ref: [SKYINFRA-321]