Add initial support for access egress external
When using access.egress.external we create a Istio serviceEntry and a netpol to be as firm as possible. Netpol has some shortcoming (i.e. not supporting dns/ fqdn), thus are limited to adresses only.
One weakness with istio serviceEntry is that the entry will be applied to the whole namespace.

Internal ref: [SKYINFRA-321]