Fix support for ports in egress external netpol
By construction each element in the egress array as JSON-objects.
This is because YAML templating wouldn't let us have a "pending" open element (i.e. we do not know if there will be either a address object, ports object or both present.) Thus we need JSON with suffix comma (which is supported syntax) to be able to support the use case where only addresses are specified, only ports are specified or both are specified.

Internal ref: [SKYINFRA-321]