Allow cidr 0.0.0.0/0 if access.egress.external has hosts rule defined
NetworkPolicy can not resolve the IP if hosts are used when defining egress external rules. Thus we need to open up the netpol to the whole internet... Suboptimal, but need to be done this way until something, like FQDNNetworkPolicies or similar are implemented.
We thus relay on the corresponding ServiceEntry to deny egress traffic that is sent to other hosts than those defined.

Internal ref: [SKYINFRA-321]